adding hosted tools
